Выполняет рефакторинг
