+++ src/const.js
@@ -1,0 +1,2 @@
+export const DESCRIPTIONS = [`Изучить теорию`, `Сделать домашку`, `Пройти интенсив на соточку`];
+export const COLORS = [`black`, `yellow`, `blue`, `green`, `pink`];
