+++ src/utils.js
@@ -1,0 +1,31 @@
+export const getRandomInteger = (min = 0, max = 1) => {
+  return Math.floor(min + Math.random() * (max - min + 1));
+};
+
+export const getRandomBoolean = () => {
+  return Boolean(getRandomInteger(0, 1));
+}
+
+export const getRandomIndex = (arr) => {
+  const randomIndex = getRandomInteger(0, arr.length - 1);
+  return arr[randomIndex];
+}
+
+export const isTaskExpired = (dueDate) => {
+  if (dueDate === null) {
+    return false;
+  }
+
+  const currentDate = new Date();
+  currentDate.setHours(23, 59, 59, 999);
+
+  return currentDate.getTime() > dueDate.getTime();
+};
+
+export const isTaskRepeating = (repeatingDays) => {
+  return Object.values(repeatingDays).some(Boolean);
+};
+
+export const humanizeTaskDueDate = (dueDate) => {
+  return dueDate.toLocaleString(`en-US`, {day: `numeric`, month: `long`});
+};
